Prevent unsafe narrowing: base/*/*
* Change to more correct types
* Add some static_casts where safe

Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3673877
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1009211}

+++ base/sampling_heap_profiler/lock_free_address_hash_set.h
@@ -84,7 +84,7 @@ class BASE_EXPORT LockFreeAddressHashSet {
   ALWAYS_INLINE Node* FindNode(void* key) const;
 
   std::vector<std::atomic<Node*>> buckets_;
-  int size_ = 0;
+  size_t size_ = 0;
   const size_t bucket_mask_;
 };
